+++ ChangeLog.md
@@ -4,7 +4,7 @@
 All notable changes to this project will be documented in this file.
 
 ## Unreleased
-
+- FIX : Ajout des lignes sur la propal créé grâce au bouton "Créer Devis" depuis un chiffrage *15/02/2022* - 1.0.1
 
 ## Version 1.0